Extract AmpAd helper
